Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/webappapis/timers/negative-settimeout-expected.txt

CONSOLE MESSAGE: Error: assert_unreached: Reached unreachable code

PASS Negative timeout in setTimeout
